Carve `TestElementwiseOps` and `TestDenseLayers` out of the monolith (wala/ML#635).
The 122 element-wise binary-operator tests (the `testAdd*` family) and
17 dense-layer tests move verbatim into their classes; the monolith
keeps 414, preserving the 1,040 total.
