+++ com.ibm.wala.cast.python.ml.test/source/com/ibm/wala/cast/python/ml/test/tensorflow/v2/TestTensorflow2Model.java
@@ -23,10 +23,8 @@ import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_1_1_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_1_2_2_27_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_1_2_FLOAT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_1_2_INT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_1_3_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_1_5_FLOAT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_1_5_INT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_2246_INT64;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_2246_OBJECT;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_25000_INT64;
@@ -63,14 +61,12 @@ import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_2_2_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_2_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_3_FLOAT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_4_INT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_5_BOOL;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_5_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_5_INT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_BOOL;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_INT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_RAGGED_RAGGED_INT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_3_STRING;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_404_13_FLOAT64;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_404_FLOAT64;
@@ -83,16 +79,12 @@ import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_4_4_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_4_8_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_4_FLOAT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_4_FLOAT64;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_4_INT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_4_RAGGED_INT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_4_RAGGED_RAGGED_NONE_STRING;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_50000_1_INT64;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_50000_32_32_3_UINT8;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_5_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_5_FLOAT64;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_5_INT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_5_RAGGED_INT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_60000_28_28_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_60000_28_28_UINT8;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_60000_784_UINT8;
@@ -106,9 +98,6 @@ import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_DYNAMIC_INT64;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_INT32_UNKNOWN_SHAPE;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_NONE_2_FLOAT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_NONE_32_FLOAT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_NONE_3_FLOAT32;
-import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_NONE_4_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_UNKNOWN_SHAPE_FLOAT32;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.TENSOR_UNKNOWN_SHAPE_UNKNOWN_DTYPE;
 import static com.ibm.wala.cast.python.ml.test.tensorflow.v2.AbstractTensorTest.UNKNOWN;
@@ -2946,546 +2935,6 @@ public class TestTensorflow2Model extends AbstractTensorTest {
         checkedAtLeastOneContext);
   }
 
-  @Test
-  public void testAdd()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_add.py", "f", 1, 1, Map.of(2, Set.of(TENSOR_5_INT32)));
-  }
-
-  @Test
-  public void testAdd2()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_add2.py", "f", 1, 1, Map.of(2, Set.of(TENSOR_5_INT32)));
-  }
-
-  @Test
-  public void testAdd3()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_add3.py", "f", 1, 1, Map.of(2, Set.of(TENSOR_5_INT32)));
-  }
-
-  @Test
-  public void testAdd4()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_add4.py", "f", 1, 1, Map.of(2, Set.of(TENSOR_5_INT32)));
-  }
-
-  @Test
-  public void testAdd5()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_add5.py", "f", 1, 1, Map.of(2, Set.of(TENSOR_5_INT32)));
-  }
-
-  @Test
-  public void testAdd6()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_add6.py", "f", 1, 1, Map.of(2, Set.of(TENSOR_5_INT32)));
-  }
-
-  @Test
-  public void testAdd7()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add7.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd8()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add8.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd9()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add9.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd10()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add10.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd11()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add11.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd12()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add12.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd13()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add13.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd14()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add14.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd15()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add15.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd16()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add16.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd17()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add17.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd18()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add18.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd19()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add19.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd20()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add20.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd21()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add21.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd22()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add22.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd23()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add23.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_INT32), 3, Set.of(TENSOR_2_INT32)));
-  }
-
-  @Test
-  public void testAdd24()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add24.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_INT32), 3, Set.of(TENSOR_2_INT32)));
-  }
-
-  @Test
-  public void testAdd25()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add25.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_INT32), 3, Set.of(TENSOR_2_INT32)));
-  }
-
-  @Test
-  public void testAdd26()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add26.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd27()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add27.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd28()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add28.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_4_INT32.asSparse()), 3, Set.of(TENSOR_3_4_INT32.asSparse())));
-  }
-
-  @Test
-  public void testAdd29()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add29.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_4_INT32.asSparse()), 3, Set.of(TENSOR_3_4_INT32.asSparse())));
-  }
-
-  @Test
-  public void testAdd30()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add30.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_INT32), 3, Set.of(TENSOR_2_2_INT32)));
-  }
-
-  @Test
-  public void testAdd31()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add31.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_INT32), 3, Set.of(TENSOR_2_2_INT32)));
-  }
-
-  @Test
-  public void testAdd32()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add32.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_INT32), 3, Set.of(TENSOR_2_INT32)));
-  }
-
-  @Test
-  public void testAdd33()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add33.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_INT32), 3, Set.of(TENSOR_2_INT32)));
-  }
-
-  @Test
-  public void testAdd34()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add34.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_3_FLOAT32), 3, Set.of(TENSOR_3_3_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd35()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add35.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_3_FLOAT32), 3, Set.of(TENSOR_3_3_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd36()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add36.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(SCALAR_TENSOR_OF_INT32), 3, Set.of(SCALAR_TENSOR_OF_INT32)));
-  }
-
-  @Test
-  public void testAdd37()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add37.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(SCALAR_TENSOR_OF_INT32), 3, Set.of(SCALAR_TENSOR_OF_INT32)));
-  }
-
-  @Test
-  public void testAdd38()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add38.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_INT32), 3, Set.of(TENSOR_5_INT32)));
-  }
-
-  @Test
-  public void testAdd39()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add39.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_INT32), 3, Set.of(TENSOR_5_INT32)));
-  }
-
-  @Test
-  public void testAdd40()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_add40.py", "func2", 1, 1, Map.of(2, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd41()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_add41.py", "func2", 1, 1, Map.of(2, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd42()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_add42.py", "func2", 1, 1, Map.of(2, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd43()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add43.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_3_FLOAT32), 3, Set.of(TENSOR_2_3_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd44()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add44.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_3_FLOAT32), 3, Set.of(TENSOR_2_3_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd45()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add45.py",
-        "value_index",
-        2,
-        2,
-        Map.of(2, Set.of(SCALAR_TENSOR_OF_FLOAT32), 3, Set.of(SCALAR_TENSOR_OF_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd46()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add46.py",
-        "value_index",
-        2,
-        2,
-        Map.of(2, Set.of(SCALAR_TENSOR_OF_FLOAT32), 3, Set.of(SCALAR_TENSOR_OF_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd47()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add47.py",
-        "add",
-        2,
-        2,
-        Map.of(2, Set.of(TENSOR_3_4_INT32.asSparse()), 3, Set.of(TENSOR_3_4_INT32.asSparse())));
-  }
-
-  @Test
-  public void testAdd48()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add48.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd49()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    // NOTE: Set the expected number of tensor variables to 3 once
-    // https://github.com/wala/ML/issues/135 is fixed.
-    test(
-        "tf2_test_add49.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd50()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add50.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_INT32), 3, Set.of(TENSOR_5_INT32)));
-  }
-
-  @Test
-  public void testAdd51()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add51.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd52()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add52.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd53()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add53.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
   /**
    * Regression guard for <a href="https://github.com/wala/ML/issues/618">wala/ML#618</a>: a custom
    * {@code fit} iterates an {@code experimental_distribute_dataset}-wrapped {@code tf.data} dataset
@@ -3906,544 +3355,6 @@ public class TestTensorflow2Model extends AbstractTensorTest {
   }
 
   @Test
-  public void testAdd54()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add54.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_NONE_32_FLOAT32), 3, Set.of(TENSOR_NONE_32_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd55()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add55.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_NONE_32_FLOAT32), 3, Set.of(TENSOR_NONE_32_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd56()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add56.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_NONE_32_FLOAT32), 3, Set.of(TENSOR_NONE_32_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd57()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add57.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_NONE_32_FLOAT32), 3, Set.of(TENSOR_NONE_32_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd58()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add58.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_NONE_32_FLOAT32), 3, Set.of(TENSOR_NONE_32_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd59()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add59.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_INT32), 3, Set.of(TENSOR_2_INT32)));
-  }
-
-  @Test
-  public void testAdd60()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add60.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_INT32), 3, Set.of(TENSOR_2_INT32)));
-  }
-
-  @Test
-  public void testAdd61()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add61.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_INT32), 3, Set.of(TENSOR_2_INT32)));
-  }
-
-  @Test
-  public void testAdd62()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add62.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_INT32), 3, Set.of(TENSOR_2_INT32)));
-  }
-
-  @Test
-  public void testAdd63()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add63.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_4_INT32.asSparse()), 3, Set.of(TENSOR_3_4_INT32.asSparse())));
-  }
-
-  @Test
-  public void testAdd64()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add64.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_4_INT32.asSparse()), 3, Set.of(TENSOR_3_4_INT32.asSparse())));
-  }
-
-  @Test
-  public void testAdd65()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add65.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_4_INT32.asSparse()), 3, Set.of(TENSOR_3_4_INT32.asSparse())));
-  }
-
-  @Test
-  public void testAdd66()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add66.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_5_INT32), 3, Set.of(TENSOR_1_5_INT32)));
-  }
-
-  @Test
-  public void testAdd67()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add67.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_5_INT32), 3, Set.of(TENSOR_1_5_INT32)));
-  }
-
-  @Test
-  public void testAdd68()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add68.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_5_INT32), 3, Set.of(TENSOR_1_5_INT32)));
-  }
-
-  @Test
-  public void testAdd69()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add69.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_5_INT32), 3, Set.of(TENSOR_1_5_INT32)));
-  }
-
-  @Test
-  public void testAdd70()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add70.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_5_INT32), 3, Set.of(TENSOR_1_5_INT32)));
-  }
-
-  @Test
-  public void testAdd71()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add71.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_5_INT32), 3, Set.of(TENSOR_1_5_INT32)));
-  }
-
-  @Test
-  public void testAdd72()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add72.py",
-        "add",
-        2,
-        3,
-        Map.of(
-            2,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING),
-            3,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING)));
-  }
-
-  @Test
-  public void testAdd73()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add73.py",
-        "add",
-        2,
-        3,
-        Map.of(
-            2,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING),
-            3,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING)));
-  }
-
-  @Test
-  public void testAdd74()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add74.py",
-        "add",
-        2,
-        3,
-        Map.of(
-            2,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING),
-            3,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING)));
-  }
-
-  @Test
-  public void testAdd75()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add75.py",
-        "add",
-        2,
-        3,
-        Map.of(
-            2,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING),
-            3,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING)));
-  }
-
-  @Test
-  public void testAdd76()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add76.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_RAGGED_RAGGED_INT32), 3, Set.of(TENSOR_3_RAGGED_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd77()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add77.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_RAGGED_RAGGED_INT32), 3, Set.of(TENSOR_3_RAGGED_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd78()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add78.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_RAGGED_RAGGED_INT32), 3, Set.of(TENSOR_3_RAGGED_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd79()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add79.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_RAGGED_RAGGED_INT32), 3, Set.of(TENSOR_3_RAGGED_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd80()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add80.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_RAGGED_RAGGED_INT32), 3, Set.of(TENSOR_3_RAGGED_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd81()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add81.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_RAGGED_RAGGED_INT32), 3, Set.of(TENSOR_3_RAGGED_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd82()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add82.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_RAGGED_RAGGED_INT32), 3, Set.of(TENSOR_3_RAGGED_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd83()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add83.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_3_RAGGED_RAGGED_INT32), 3, Set.of(TENSOR_3_RAGGED_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd84()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add84.py",
-        "add",
-        2,
-        3,
-        Map.of(
-            2,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING),
-            3,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING)));
-  }
-
-  @Test
-  public void testAdd85()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add85.py",
-        "add",
-        2,
-        3,
-        Map.of(
-            2,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING),
-            3,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING)));
-  }
-
-  @Test
-  public void testAdd86()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add86.py",
-        "add",
-        2,
-        3,
-        Map.of(
-            2,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING),
-            3,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING)));
-  }
-
-  @Test
-  public void testAdd87()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add87.py",
-        "add",
-        2,
-        3,
-        Map.of(
-            2,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING),
-            3,
-            Set.of(TENSOR_4_RAGGED_RAGGED_NONE_STRING)));
-  }
-
-  @Test
-  public void testAdd88()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add88.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_RAGGED_INT32), 3, Set.of(TENSOR_5_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd89()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add89.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_RAGGED_INT32), 3, Set.of(TENSOR_5_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd90()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add90.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_RAGGED_INT32), 3, Set.of(TENSOR_5_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd91()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add91.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_RAGGED_INT32), 3, Set.of(TENSOR_5_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd92()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add92.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_RAGGED_INT32), 3, Set.of(TENSOR_5_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd93()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add93.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_RAGGED_INT32), 3, Set.of(TENSOR_5_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd94()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add94.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_RAGGED_INT32), 3, Set.of(TENSOR_5_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd95()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add95.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_RAGGED_INT32), 3, Set.of(TENSOR_5_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd96()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add96.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_5_RAGGED_INT32), 3, Set.of(TENSOR_5_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd97()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add97.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_4_RAGGED_INT32), 3, Set.of(TENSOR_4_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd98()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add98.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_4_RAGGED_INT32), 3, Set.of(TENSOR_4_RAGGED_INT32)));
-  }
-
-  @Test
-  public void testAdd99()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add99.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_4_RAGGED_INT32), 3, Set.of(TENSOR_4_RAGGED_INT32)));
-  }
-
-  @Test
   public void testStrictnessFailure()
       throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
     test(
@@ -4458,224 +3369,6 @@ public class TestTensorflow2Model extends AbstractTensorTest {
   public void testNoneDType()
       throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
     test("tf2_test_none_dtype.py", "test_none_dtype", 1, 1, Map.of(2, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd100()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add100.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_10_2_FLOAT32), 3, Set.of(TENSOR_10_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd101()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add101.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_10_2_FLOAT32), 3, Set.of(TENSOR_10_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd102()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add102.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_10_2_FLOAT32), 3, Set.of(TENSOR_10_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd103()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add103.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_10_2_FLOAT32), 3, Set.of(TENSOR_10_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd104()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add104.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_4_FLOAT32), 3, Set.of(TENSOR_4_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd105()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add105.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_4_FLOAT32), 3, Set.of(TENSOR_4_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd106()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add106.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_4_FLOAT32), 3, Set.of(TENSOR_4_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd107()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add107.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_10_2_FLOAT32), 3, Set.of(TENSOR_10_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd108()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add108.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_10_2_FLOAT32), 3, Set.of(TENSOR_10_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd109()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add109.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd110()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add110.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd111()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add111.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_3_FLOAT32.asSparse()), 3, Set.of(TENSOR_2_3_FLOAT32.asSparse())));
-  }
-
-  @Test
-  public void testAdd112()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add112.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_3_FLOAT32.asSparse()), 3, Set.of(TENSOR_2_3_FLOAT32.asSparse())));
-  }
-
-  @Test
-  public void testAdd113()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add113.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_3_FLOAT32.asSparse()), 3, Set.of(TENSOR_2_3_FLOAT32.asSparse())));
-  }
-
-  @Test
-  public void testAdd114()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add114.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_NONE_32_FLOAT32), 3, Set.of(TENSOR_NONE_32_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd115()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add115.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_NONE_32_FLOAT32), 3, Set.of(TENSOR_NONE_32_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd116()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add116.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  /**
-   * Tests that under per-context shape unions the analysis returns the broadcastable cross-pairs
-   * and silently discards the non-broadcastable ones (wala/ML#462).
-   *
-   * <p>In {@code tf2_test_add117.py}, the variable {@code a} can be either 1 or 3.
-   *
-   * <ul>
-   *   <li>If {@code a=1}, the addition is {@code [1, 2] + [2, 2]}, which is broadcastable to {@code
-   *       [2, 2]}.
-   *   <li>If {@code a=3}, the addition is {@code [3, 2] + [2, 2]}, which is NOT broadcastable.
-   * </ul>
-   *
-   * The analysis retains the broadcastable result ({@code [2, 2]}) and discards the
-   * non-broadcastable cross-pair as analysis-level imprecision rather than a runtime error &mdash;
-   * the cross-pair would never co-occur at runtime under matched contexts. When <em>every</em> pair
-   * is non-broadcastable, the result shape instead degrades to ⊤ (<a
-   * href="https://github.com/wala/ML/issues/583">wala/ML#583</a>).
-   *
-   * @see #testAdd117a()
-   */
-  @Test
-  public void testAdd117()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add117.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32, TENSOR_3_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
   }
 
   /**
@@ -4708,17 +3401,6 @@ public class TestTensorflow2Model extends AbstractTensorTest {
   }
 
   @Test
-  public void testAdd118()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add118.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_INT32), 3, Set.of(TENSOR_2_2_INT32)));
-  }
-
-  @Test
   public void testAddResultFlow()
       throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
     test("tf2_test_add_result_flow.py", "check_result", 1, 1, Map.of(2, Set.of(TENSOR_2_FLOAT32)));
@@ -4740,50 +3422,6 @@ public class TestTensorflow2Model extends AbstractTensorTest {
   public void testDivResultFlow()
       throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
     test("tf2_test_div_result_flow.py", "check_result", 1, 1, Map.of(2, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd119()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add119.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_2_FLOAT32), 3, Set.of(TENSOR_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd120()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add120.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_1_2_FLOAT32), 3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd121()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add121.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_4_FLOAT32), 3, Set.of(TENSOR_4_FLOAT32)));
-  }
-
-  @Test
-  public void testAdd122()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_add122.py",
-        "add",
-        2,
-        3,
-        Map.of(2, Set.of(TENSOR_4_FLOAT64), 3, Set.of(TENSOR_4_FLOAT64)));
   }
 
   /**
@@ -8626,84 +7264,6 @@ public class TestTensorflow2Model extends AbstractTensorTest {
   }
 
   /**
-   * NLPGNN's {@code DenseLayer3d} einsum path in miniature (wala/ML#704): the weight is built flat
-   * from configuration fields, reshaped to rank 3 in {@code call} (the {@code hidden} leading dim
-   * stays dynamic since {@code build}'s {@code input_shape} subscript doesn't resolve), and
-   * consumed by {@code einsum("BFH,HND->BFND", ...)}. The parser refines the contracted {@code H}
-   * label's dynamic occurrence with the input's known {@code 6} and composes the precise runtime
-   * {@code (2, 4, 3, 5)}: the trailing head dims are static, per the issue's expectation.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dEinsum()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_einsum.py",
-        "consume",
-        1,
-        1,
-        Map.of(2, Set.of(TensorType.of(FLOAT_32, 2, 4, 3, 5))));
-  }
-
-  /**
-   * NLPGNN's {@code DenseLayer3dProj} (wala/ML#704): the input arrives with an unresolvable shape
-   * (the wala/ML#711 opaque-bound reshape), but {@code einsum("BFND,NDH->BFH", input_tensor, w)}
-   * proves the input is rank 4, and the shared {@code N}/{@code D} labels take their extents from
-   * the reshaped weight's statically-known {@code (3, 5, 6)}. The {@code B}/{@code F} axes carry no
-   * {@code None} evidence, so they are {@code Unresolved} (wala/ML#721), not {@code Dynamic}.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dProj()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_proj.py",
-        "DenseLayer3dProj.call",
-        1,
-        4,
-        Map.of(
-            3,
-            Set.of(
-                new TensorType(
-                    FLOAT_32,
-                    asList(
-                        UnresolvedDim.INSTANCE,
-                        UnresolvedDim.INSTANCE,
-                        new NumericDim(3),
-                        new NumericDim(5))))));
-  }
-
-  /**
-   * Operand-order companion of {@link #testDense3dEinsum()} (wala/ML#704): the weight, whose
-   * contracted dim is dynamic, comes first in the equation ({@code "HND,BFH->BFND"}), so the
-   * input's statically-known occurrence of the shared {@code H} label arrives second and refines
-   * the earlier dynamic one. The composed shape is the same precise {@code (2, 4, 3, 5)}.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dEinsum2()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_einsum2.py",
-        "consume",
-        1,
-        1,
-        Map.of(2, Set.of(TensorType.of(FLOAT_32, 2, 4, 3, 5))));
-  }
-
-  /**
    * Distilled regression guard for the scalar-expression broadcast (wala/ML#718): a tensor scaled
    * by a statically opaque scalar expression ({@code 1.0 / math.sqrt(4.0)}) keeps its shape, since
    * the expression's scalarness is structural even though its value never resolves. The analysis
@@ -8797,227 +7357,6 @@ public class TestTensorflow2Model extends AbstractTensorTest {
         1,
         1,
         Map.of(2, Set.of(TensorType.of(FLOAT_32, 2, 4, 5))));
-  }
-
-  /**
-   * Composition of {@link #testDense3dEinsum()} and {@link #testEinsumViaMatmul()} (wala/ML#704):
-   * NLPGNN's {@code DenseLayer3d.call} delegates to the module-level {@code einsum_via_matmul}
-   * helper. The {@code input_tensor} parameter carries the precise type across the layer-call
-   * boundary, and the {@code w} parameter is fully static: the trailing dimensions are the folded
-   * configuration fields, the leading (hidden) dimension resolves through {@code build}'s {@code
-   * input_shape} subscript ({@code self.hidden_size = input_shape[2]}, wala/ML#712), and the
-   * literal-shape pin agrees with the generator-side result (wala/ML#713), so the union is a single
-   * member.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dMatmul()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_matmul.py",
-        "einsum_via_matmul",
-        2,
-        14,
-        Map.of(
-            2,
-            Set.of(TensorType.of(FLOAT_32, 2, 4, 6)),
-            3,
-            Set.of(TensorType.of(FLOAT_32, 6, 3, 5))));
-  }
-
-  /**
-   * Escalation of {@link #testDense3dMatmul()} to NLPGNN's real dispatch shape (wala/ML#704): the
-   * {@code DenseLayer3d} is created in an outer attention layer's {@code build}, stored as an
-   * attribute, and invoked through it in the outer {@code call} — two trampoline hops before {@code
-   * einsum_via_matmul} sees the input tensor. The {@code input_tensor} parameter carries the
-   * precise type through both hops, and the {@code w} parameter resolves exactly as in the one-hop
-   * composition, including the hidden dimension through {@code build}'s {@code input_shape}
-   * subscript (wala/ML#712).
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dMatmul2()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_matmul2.py",
-        "einsum_via_matmul",
-        2,
-        14,
-        Map.of(
-            2,
-            Set.of(TensorType.of(FLOAT_32, 2, 4, 6)),
-            3,
-            Set.of(TensorType.of(FLOAT_32, 6, 3, 5))));
-  }
-
-  /**
-   * Negative-index companion of {@link #testDense3dMatmul()} (wala/ML#712): {@code build} stores
-   * {@code input_shape[-1]}, exercising the Python negative-index arm of the shape-vector subscript
-   * resolution.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dMatmul3()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_matmul3.py",
-        "einsum_via_matmul",
-        2,
-        14,
-        Map.of(
-            2,
-            Set.of(TensorType.of(FLOAT_32, 2, 4, 6)),
-            3,
-            Set.of(TensorType.of(FLOAT_32, 6, 3, 5))));
-  }
-
-  /**
-   * Explicit-{@code build} companion of {@link #testDense3dMatmul()} (wala/ML#712): the script
-   * calls {@code layer.build(x.shape)} before the layer call, so the {@code input_shape} parameter
-   * also receives a real argument, exercising the explicit-caller arm of the walk alongside the
-   * lazy-build trampoline hop.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dMatmul4()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_matmul4.py",
-        "einsum_via_matmul",
-        2,
-        14,
-        Map.of(
-            2,
-            Set.of(TensorType.of(FLOAT_32, 2, 4, 6)),
-            3,
-            Set.of(TensorType.of(FLOAT_32, 6, 3, 5))));
-  }
-
-  /**
-   * Two-instance guard for the {@code build} subscript resolution (wala/ML#712): the class is
-   * instantiated on inputs with different hidden sizes, so the per-class attribute chase sees
-   * conflicting {@code hidden_size} stores and must bail; the weight's leading dimension stays
-   * dynamic (a sound conservative result) while the folded configuration fields keep their
-   * constants.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dMatmul5()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_matmul5.py",
-        "einsum_via_matmul",
-        2,
-        14,
-        Map.of(
-            2,
-            Set.of(TensorType.of(FLOAT_32, 2, 4, 6), TensorType.of(FLOAT_32, 2, 4, 8)),
-            3,
-            Set.of(
-                new TensorType(
-                    FLOAT_32,
-                    asList(UnresolvedDim.INSTANCE, new NumericDim(3), new NumericDim(5))))));
-  }
-
-  /**
-   * Arithmetic-over-subscript companion of {@link #testDense3dMatmul()} (wala/ML#714): {@code
-   * build} derives the head size the way NLPGNN's {@code ALBERTAttention.build} does, dividing a
-   * shape subscript by a configuration field under an {@code int(...)} wrapper ({@code
-   * self.head_size = int(input_shape[2]/self.num_attention_heads)}), so the reshaped weight's
-   * dimensions all fold.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dMatmul6()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_matmul6.py",
-        "einsum_via_matmul",
-        2,
-        14,
-        Map.of(
-            2,
-            Set.of(TensorType.of(FLOAT_32, 2, 4, 6)),
-            3,
-            Set.of(TensorType.of(FLOAT_32, 6, 3, 2))));
-  }
-
-  /**
-   * Bare-division guard companion of {@link #testDense3dMatmul6()} (wala/ML#714): a branch the
-   * runtime never takes stores {@code input_shape[2] / 4}, a bare float division the fold must
-   * refuse (its non-exact truncation is undefined without the {@code int(...)} wrapper), so the
-   * attribute is unresolvable and the weight's trailing dimension soundly stays dynamic.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dMatmul7()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_matmul7.py",
-        "einsum_via_matmul",
-        2,
-        14,
-        Map.of(
-            2,
-            Set.of(TensorType.of(FLOAT_32, 2, 4, 6)),
-            3,
-            Set.of(
-                new TensorType(
-                    FLOAT_32,
-                    asList(new NumericDim(6), new NumericDim(3), UnresolvedDim.INSTANCE)))));
-  }
-
-  /**
-   * Nested-arithmetic companion of {@link #testDense3dMatmul6()} (wala/ML#714): the divisor is
-   * itself an arithmetic expression over a configuration field ({@code
-   * int(input_shape[2]/(self.num_attention_heads - 1))}), so the operand resolution recurses.
-   *
-   * @throws ClassHierarchyException if the class hierarchy cannot be built.
-   * @throws IllegalArgumentException if the input fixture is malformed.
-   * @throws CancelException if the analysis is cancelled.
-   * @throws IOException if the input fixture cannot be read.
-   */
-  @Test
-  public void testDense3dMatmul8()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test(
-        "tf2_test_dense3d_matmul8.py",
-        "einsum_via_matmul",
-        2,
-        14,
-        Map.of(
-            2,
-            Set.of(TensorType.of(FLOAT_32, 2, 4, 6)),
-            3,
-            Set.of(TensorType.of(FLOAT_32, 6, 3, 3))));
   }
 
   /**
@@ -9380,67 +7719,6 @@ public class TestTensorflow2Model extends AbstractTensorTest {
         Map.of(
             2, Set.of(TENSOR_2_FLOAT64),
             3, Set.of(TENSOR_2_2_FLOAT32)));
-  }
-
-  @Test
-  public void testDense()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_dense.py", "consume", 1, 1, Map.of(2, Set.of(TENSOR_NONE_4_FLOAT32)));
-  }
-
-  /**
-   * Test for <a href="https://github.com/wala/ML/issues/371">wala/ML#371</a>. A single {@code
-   * Dense} layer call inside {@code M.call} with a {@code tf.keras.Input} parameter.
-   *
-   * <p>Two tensor variables are found: the {@code x} parameter (v3, shape {@code (None, 3)}) and
-   * the {@code Dense} result (v25, shape {@code (None, 4)}). Both are correct source-level tensors
-   * under a single trampoline context.
-   */
-  @Test
-  public void testDenseModelCall()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_dense_model_call.py", "M.call", 1, 2, Map.of(3, Set.of(TENSOR_NONE_3_FLOAT32)));
-  }
-
-  @Test
-  public void testDense2()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_dense2.py", "consume1", 1, 1, Map.of(2, Set.of(TENSOR_NONE_4_FLOAT32)));
-    test("tf2_test_dense2.py", "consume2", 1, 1, Map.of(2, Set.of(TENSOR_NONE_2_FLOAT32)));
-  }
-
-  /**
-   * Chained {@code Dense} layer calls at module level, where the second layer's {@code inputs}
-   * argument is the return value of the first layer's call. Exercises shape propagation through a
-   * layer-call result at script-body scope.
-   */
-  @Test
-  public void testDenseChain()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_dense_chain.py", "consume", 1, 1, Map.of(2, Set.of(TENSOR_NONE_2_FLOAT32)));
-  }
-
-  /**
-   * Chained {@code Dense} layer calls inside a {@code tf.keras.Model.__call__} method body with
-   * direct {@code self.layer1} / {@code self.layer2} attribute reads. Exercises shape propagation
-   * through a layer-call result inside a user-defined class method.
-   */
-  @Test
-  public void testDenseChain2()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_dense_chain2.py", "consume", 1, 1, Map.of(2, Set.of(TENSOR_NONE_2_FLOAT32)));
-  }
-
-  /**
-   * Chained {@code Dense} layer calls inside a {@code tf.keras.Model.__call__} method body, where
-   * the layers are iterated via a {@code for} loop over a {@code self.layers_list} attribute rather
-   * than being accessed by direct attribute name. Exercises shape propagation through a loop-phi'd
-   * local whose points-to set spans every list element.
-   */
-  @Test
-  public void testDenseChain3()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    test("tf2_test_dense_chain3.py", "consume", 1, 1, Map.of(2, Set.of(TENSOR_NONE_4_FLOAT32)));
   }
 
   /**
